backend/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
groq>=1.1.0
supabase>=2.28.0
pypdf==4.0.0
python-docx==1.1.0
requests==2.31.0
PyJWT==2.8.0
cryptography==41.0.3
